pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel>=0.41"]
build-backend = "setuptools.build_meta"

[project]
name = "envguard"
version = "0.1.0"
description = "Self-healing dev environment doctor"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "rich>=13.7.1",
  "PyYAML>=6.0.1",
]

[project.scripts]
envguard = "envguard.cmd.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100